at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil

+++ os/win_graphics_path.cpp
@@ -206,7 +206,7 @@ namespace win
    bool graphics_path::set(const ::ca::graphics_path::move & p)
    {
 
-      internal_add_move(p.m_x, p.m_y);
+      internal_add_move((int32_t) p.m_x, (int32_t) p.m_y);
 
 
       return true;
